Split up readers and from_reader to separate public modules
