+++ barse/src/reader/const_endian_byte_reader.rs
